pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "jezr"
version = "0.1.0"
description = "Training intelligence for endurance athletes"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "python-dateutil",
    "anthropic",
]

[project.scripts]
jezr = "jezr.cli:main"

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["."]
include = ["jezr*"]
